Replace Math.random() with SHA256 digests for candidate quality values (#15)
There's no significant reason to do this --Math.random() would work just fine for the scope of this simulation-- but hash functions are fun and this guarantees no two quality values with be the same. This is also set up work for later adding the functionality to display and optionally set the initial seed hash value.